Move private key into App. Introduce algorithmic session provider and move relevant methods there.
