+++ seacatauth/cookie/handler.py
@@ -445,7 +445,7 @@ class CookieHandler(object):
 		# First try if the cookie is a JWToken
 		if "." in cookie_value:
 			try:
-				return await self.CookieService.OpenIdConnectService.build_algorithmic_session_from_token(cookie_value)
+				return await self.SessionService.Algorithmic.build_algorithmic_session_from_token(cookie_value)
 			except asab.exceptions.NotAuthenticatedError:
 				# The JWToken is invalid or expired
 				return None
